Split larch/state/stall_recovery.py (2,905 LOC) into cohesive sibling modules
Decompose the god-module into 9 underscore-prefixed private modules under
python/larch/state/, each grouped by responsibility:
  _tokens.py      - token allowlists, safe-value rendering, shared utilities
  _detail_log.py  - failure detail log validation and reading
  _escalation.py  - escalation artifact utilities and recording
  _validate.py    - terminal state and token validation
  _classify.py    - stall classification and attempt tracking
  _normalize.py   - outcome normalization and issue-env normalization
  _corpus.py      - sensitive corpus building and tier-B validation
  _state_mgmt.py  - state file management (clear/seed stall state)
  _report.py      - report composition, deduplication, chat printing

stall_recovery.py is reduced to CLI wiring and re-exports. No behavior change.
Public *_main entry points, wire formats, and CLI contract are preserved.
Updates complexity-baseline.json, ruff.toml, and test_stall_recovery.py.
